Add files via upload
I built this project to create a task sheet for the day.
We can enroll tasks for today.  we will insert it in the remaining task table. if I have done that task then I need to click on the check box so it will remove it from the remaining table and shift to the Done table.
For the backend database, I used crudcrud via Axios API calls